Make changes
Change in-memory for db, authentication and jwt_required() in places, reviews, users api

+++ part3/hbnb/app/api/v1/reviews.py
@@ -1,5 +1,7 @@
 from flask_restx import Namespace, Resource, fields, _http
 from app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity
+
 
 api = Namespace('reviews', description='Review operations')
 
@@ -40,10 +42,46 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @api.response(404, 'Place not found')
+    @jwt_required()
     def post(self):
         """Register a new review"""
+        current_user = get_jwt_identity()
         # Automatically parses and validates request JSON
         review_data = api.payload
+        # Retrieve the place_id
+        place_id_to_review = review_data.get('place_id')
+        if not place_id_to_review:
+            api.abort(400, error='Missing place_id')
+            return {'error': 'Missing place_id'}, 400
+        # Retrieve the place to validate ownership
+        place_to_review = facade.get_place(place_id_to_review)
+        if not place_to_review:
+            api.abort(404, error=str(e))
+            return {'error': str(e)}, 404
+
+        # Check if the current user tries to review their own place
+        if place_to_review.owner_id == current_user['id']:
+            api.abort(403, error='You cannot review your own place')
+            return {'error': 'You cannot review your own place'}, 403
+
+        # Check if the review already exists
+        existing_review = facade.get_review_by_place_and_user(
+            place_id_to_review, current_user['id'])
+        if existing_review:
+            api.abort(403, error='You have already reviewed this place')
+            return {'error': 'You have already reviewed this place'}, 403
+
+        # Ensure user_id in payload matches authenticated user
+        if 'user_id' in review_data and \
+                review_data['user_id'] != current_user['id']:
+            api.abort(403, error='You are not allowed to create a review for'
+                      'another user')
+            return {'error': 'You are not allowed to create a review for'
+                    'another user'}, 403
+
+        # Set user_id to the authenticated user's ID
+        review_data['user_id'] = current_user['id']
         try:
             review = facade.create_review(review_data)
         except Exception as e:
@@ -92,9 +130,26 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(400, 'Invalid input data')
     @api.response(404, 'Review not found')
+    @api.response(403, 'Unauthorized action')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
+        current_user = get_jwt_identity()
+        # Retrieve the review to validate ownership
+        review_to_update = facade.get_review(review_id)
+
+        # Check if the review exists
+        if not review_to_update:
+            api.abort(404, error=str(e))
+            return {'error': str(e)}, 404
+
+        # Check if the authenticated user is the creator of the review
+        if review_to_update.user_id != current_user['id']:
+            api.abort(403, error=str(e))
+            return {'error': str(e)}, 403
+
         review_data = api.payload
+
         try:
             updated_review = facade.update_review(review_id,
                                                   review_data)
@@ -110,12 +165,28 @@ class ReviewResource(Resource):
     @api.doc('Deletes review')
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @api.response(403, 'Unauthorized action')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
+        current_user = get_jwt_identity()
+        # Retrieve the review to validate ownership
+        review_to_delete = facade.get_review(review_id)
+
+        # Check if the review already exists
+        if not review_to_delete:
+            api.abort(404, error=str(e))
+            return {'error': str(e)}, 404
+
+        # Check if the creator of the review is the current user
+        if review_to_delete.user_id != current_user['id']:
+            api.abort(403, error=str(e))
+            return {'error': str(e)}, 403
+
         try:
             facade.delete_review(review_id)
         except Exception as e:
-            api.abort(404, str(e))
+            api.abort(404, error=str(e))
             return {'error': str(e)}, 404
         return f"Review {review_id} has been succesfully deleted", 200
